chore: remove unused Db.Util functions
Also, moving Node-related functions to Node.Db, and out of Db.Util
