+++ lib/ae_mdw/db/util.ex
@@ -1,18 +1,17 @@
 defmodule AeMdw.Db.Util do
-  # credo:disable-for-this-file
+  @moduledoc false
+
   alias AeMdw.Blocks
   alias AeMdw.Collection
   alias AeMdw.Db.Model
   alias AeMdw.Db.State
   alias AeMdw.Txs
 
-  require Logger
   require Model
-
-  import AeMdw.Util
 
   @typep state() :: State.t()
 
+  @spec read_tx!(state(), Txs.txi()) :: Model.tx()
   def read_tx!(state, txi), do: State.fetch!(state, Model.Tx, txi)
 
   @spec read_block!(state(), Blocks.block_index()) :: Model.block()
@@ -28,107 +27,12 @@ defmodule AeMdw.Db.Util do
     txi
   end
 
+  @spec last_gen(state()) :: Blocks.height()
   def last_gen(state) do
     case State.prev(state, Model.Block, nil) do
       {:ok, {height, _mbi}} -> height
       :none -> raise RuntimeError, message: "can't get last key for table Model.Block"
     end
-  end
-
-  def ensure_key!(tab, getter) do
-    case apply(__MODULE__, getter, [tab]) do
-      :"$end_of_table" ->
-        raise RuntimeError, message: "can't get #{getter} key for table #{tab}"
-
-      k ->
-        k
-    end
-  end
-
-  def collect_keys(tab, acc, start_key, next_fn, progress_fn) do
-    case next_fn.(tab, start_key) do
-      :"$end_of_table" ->
-        acc
-
-      next_key ->
-        case progress_fn.(next_key, acc) do
-          {:halt, res_acc} -> res_acc
-          {:cont, next_acc} -> collect_keys(tab, next_acc, next_key, next_fn, progress_fn)
-        end
-    end
-  end
-
-  def do_writes(tab_xs, db_write) when is_function(db_write, 2),
-    do: Enum.each(tab_xs, fn {tab, xs} -> Enum.each(xs, &db_write.(tab, &1)) end)
-
-  def tx_val(tx_rec, field),
-    do: tx_val(tx_rec, elem(tx_rec, 0), field)
-
-  def tx_val(tx_rec, tx_type, field),
-    do: elem(tx_rec, Enum.find_index(AeMdw.Node.tx_fields(tx_type), &(&1 == field)) + 1)
-
-  def gen_collect(table, init_key_probe, key_tester, progress, new, add, return) do
-    return.(
-      case progress.(table, init_key_probe) do
-        :"$end_of_table" ->
-          new.()
-
-        start_key ->
-          case key_tester.(start_key) do
-            false ->
-              new.()
-
-            other ->
-              init_acc =
-                case other do
-                  :skip -> new.()
-                  true -> add.(start_key, new.())
-                end
-
-              collect_keys(table, init_acc, start_key, progress, fn key, acc ->
-                case key_tester.(key) do
-                  false -> {:halt, acc}
-                  true -> {:cont, add.(key, acc)}
-                  :skip -> {:cont, acc}
-                end
-              end)
-          end
-      end
-    )
-  end
-
-  ##########
-
-  def msecs(msecs) when is_integer(msecs) and msecs > 0, do: msecs
-  def msecs(%Date{} = d), do: msecs(date_time(d))
-  def msecs(%DateTime{} = d), do: DateTime.to_unix(d) * 1000
-
-  def date_time(%DateTime{} = dt),
-    do: dt
-
-  def date_time(msecs) when is_integer(msecs) and msecs > 0,
-    do: DateTime.from_unix(div(msecs, 1000)) |> ok!
-
-  def date_time(%Date{} = d) do
-    {:ok, dt, 0} = DateTime.from_iso8601(Date.to_iso8601(d) <> " 00:00:00.0Z")
-    dt
-  end
-
-  def prev_block_type(header) do
-    prev_hash = :aec_headers.prev_hash(header)
-    prev_key_hash = :aec_headers.prev_key_hash(header)
-
-    cond do
-      :aec_headers.height(header) == 0 -> :key
-      prev_hash == prev_key_hash -> :key
-      true -> :micro
-    end
-  end
-
-  def proto_vsn(height) do
-    hps = AeMdw.Node.height_proto()
-    [{vsn, _} | _] = Enum.drop_while(hps, fn {_vsn, min_h} -> height < min_h end)
-    vsn
   end
 
   @spec block_txi(state(), Blocks.block_index()) :: Txs.txi() | nil
